이클립스 버전 업데이트에 따른 변경 사항

+++ 백준 알고리즘/src/DP/Problem_1256.java
@@ -18,7 +18,7 @@ public class Problem_1256 {
 		int N = Integer.parseInt(stk.nextToken());
 		int M = Integer.parseInt(stk.nextToken());
 		
-		long K = (long)Integer.parseInt(stk.nextToken());
+		long K = Integer.parseInt(stk.nextToken());
 		
 		
 		int n = N + M;
